Added 32bit versions of libraries.

+++ lib/definitions.h
@@ -22,8 +22,10 @@
 #define _DEFINITIONS_H_
 
 // Constants
-typedef long long LONG;
-typedef unsigned long long ULONG;
+// typedef long long LONG;
+// typedef unsigned long long ULONG;
+typedef int LONG;
+typedef unsigned int ULONG;
 typedef int INT;
 typedef unsigned int UINT;
 typedef unsigned long long NodeID;
